Update default model behavior
Previous behavior would use the default-value of a property's type on the model when no binding took place on the property.
However, this negates a model that has set a default-value for the property.

New behavior is a temporary shim to get the expected behavior while leaving a breadcrumb for an idea to compare the "hydrated" model against a plain "new-instance" of the model.