Add bind support for collections
New behavior does additional checks on data coming from value providers. There are now two possibilities: the provider returns one-value or a collection of values.
The former does what is already being done.
The latter goes through an additional process to take the collection, convert each element, and insert into a temporary list matching the type of the model's property.
Assuming counts from the value provider's collection and the temporary list match, then the values are set on the model.